DTO를 다시 개별 클래스로 분리
