+++ backend/tiggle/src/main/java/com/side/tiggle/domain/comment/dto/req/CommentCreateReqDto.java
@@ -1,0 +1,9 @@
+package com.side.tiggle.domain.comment.dto.req;
+
+import com.side.tiggle.domain.comment.dto.CommentDto;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CommentCreateReqDto extends CommentDto{}
