+++ backend/tiggle/src/main/java/com/side/tiggle/domain/comment/dto/req/CommentUpdateReqDto.java
@@ -1,0 +1,10 @@
+package com.side.tiggle.domain.comment.dto.req;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CommentUpdateReqDto {
+    String content;
+}
