+++ backend/tiggle/src/test/java/com/side/tiggle/domain/comment/CommentServiceTest.java
@@ -1,5 +1,7 @@
 package com.side.tiggle.domain.comment;
 
+import com.side.tiggle.domain.comment.dto.req.CommentCreateReqDto;
+import com.side.tiggle.domain.comment.dto.req.CommentUpdateReqDto;
 import com.side.tiggle.domain.comment.model.Comment;
 import com.side.tiggle.domain.comment.repository.CommentRepository;
 import com.side.tiggle.domain.comment.service.CommentService;
@@ -9,7 +11,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
@@ -26,7 +27,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("사용자는 코멘트 작성에 성공한다")
     void 코멘트_작성() {
-        CommentDto dto = new CommentDto();
+        CommentCreateReqDto dto = new CommentCreateReqDto();
         dto.setSenderId(1L);
         dto.setReceiverId(2L);
         dto.setTxId(1L);
@@ -41,7 +42,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("유효하지 않은 tx에 코멘트를 작성할 수 없다")
     void 코멘트_작성_실패1() {
-        CommentDto dto = new CommentDto();
+        CommentCreateReqDto dto = new CommentCreateReqDto();
         dto.setSenderId(1L);
         dto.setReceiverId(2L);
         dto.setTxId(0L);
@@ -55,7 +56,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("유효하지 않은 parentId로 코멘트를 작성할 수 없다")
     void 코멘트_작성_실패2() {
-        CommentDto dto = new CommentDto();
+        CommentCreateReqDto dto = new CommentCreateReqDto();
         dto.setSenderId(1L);
         dto.setReceiverId(2L);
         dto.setTxId(1L);
@@ -70,7 +71,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("유효하지 않은 receiverId로 코멘트를 작성할 수 없다")
     void 코멘트_작성_실패3() {
-        CommentDto dto = new CommentDto();
+        CommentCreateReqDto dto = new CommentCreateReqDto();
         dto.setSenderId(1L);
         dto.setReceiverId(0L);
         dto.setTxId(1L);
@@ -84,7 +85,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("사용자는 코멘트 수정에 성공한다")
     void 코멘트_수정(){
-        CommentDto.Request.Update dto = new CommentDto.Request.Update();
+        CommentUpdateReqDto dto = new CommentUpdateReqDto();
         dto.setContent("Updated comment");
 
         Long commentId = 1L;
@@ -100,7 +101,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("사용자는 자신의 것이 아닌 코멘트를 수정할 수 없다")
     void 코멘트_수정_실패1() {
-        CommentDto.Request.Update dto = new CommentDto.Request.Update();
+        CommentUpdateReqDto dto = new CommentUpdateReqDto();
         dto.setContent("Updated comment");
 
         Long commentId = 1L;
@@ -116,7 +117,7 @@ public class CommentServiceTest {
 
         // TODO 이 테스트는 실패한다
         Long senderId = 3L;
-        CommentDto dto = new CommentDto();
+        CommentCreateReqDto dto = new CommentCreateReqDto();
         dto.setSenderId(senderId);
         dto.setReceiverId(2L);
         dto.setTxId(1L);
@@ -134,7 +135,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("사용자는 자신의 것이 아닌 코멘트를 삭제할 수 없다")
     void 코멘트_삭제_실패1() {
-        CommentDto.Request.Update dto = new CommentDto.Request.Update();
+        CommentUpdateReqDto dto = new CommentUpdateReqDto();
         dto.setContent("Updated comment");
 
         Long commentId = 1L;
@@ -147,7 +148,7 @@ public class CommentServiceTest {
     @Test
     @DisplayName("대댓글 생성과 조회를 성공한다")
     void 대댓글_생성과_조회_성공() {
-        CommentDto dto = new CommentDto();
+        CommentCreateReqDto dto = new CommentCreateReqDto();
         dto.setContent("대댓글 추가");
         dto.setParentId(1L);
         dto.setSenderId(1L);
@@ -156,7 +157,7 @@ public class CommentServiceTest {
 
         commentService.createComment(dto);
 
-        Page<Comment> comment = commentService.getChildrenByParentId(dto.parentId, 0, Integer.MAX_VALUE );
+        Page<Comment> comment = commentService.getChildrenByParentId(dto.getParentId(), 0, Integer.MAX_VALUE );
         Assertions.assertFalse(comment.isEmpty());
         Assertions.assertFalse(comment.getContent().isEmpty());
     }
@@ -165,13 +166,13 @@ public class CommentServiceTest {
     @DisplayName("트랜잭션의 댓글을 조회한다")
     void 트랜잭션_댓글_조회_성공(){
         Long txId = 1L;
-        CommentDto dto = new CommentDto();
+        CommentCreateReqDto dto = new CommentCreateReqDto();
         dto.setContent("댓글 추가");
         dto.setSenderId(1L);
         dto.setReceiverId(2L);
         dto.setTxId(txId);
 
-        CommentDto dto1 = new CommentDto();
+        CommentCreateReqDto dto1 = new CommentCreateReqDto();
         dto1.setContent("대댓글 추가");
         dto1.setParentId(1L);
         dto1.setSenderId(1L);
@@ -185,12 +186,12 @@ public class CommentServiceTest {
 
         // 댓글을 반환한다
         Optional<Comment> shouldReturn = comments.getContent().stream().filter(it ->
-                it.getContent().equals(dto.content)).findFirst();
+                it.getContent().equals(dto.getContent())).findFirst();
         Assertions.assertTrue(shouldReturn.isPresent());
 
         // 대댓글은 반환하지 않는다
         Optional<Comment> shouldNotReturn = comments.getContent().stream().filter(it ->
-                it.getContent().equals(dto1.content)).findFirst();
+                it.getContent().equals(dto1.getContent())).findFirst();
         Assertions.assertTrue(shouldNotReturn.isEmpty());
     }
